fastapi-poc/pyproject.toml
[tool.poetry]
name = "barc"
version = "0.1.0"
description = ""
authors = ["Glayson Olivieri"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.1"
python-dotenv = "^1.0.1"
psycopg = "^3.1.19"
pyjwt = "^2.8.0"
argon2-cffi = "^23.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
bpython = "^0.24"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
